Modify shader to work on absolute coordinate
Previous implementation was taking the coordinate system in normalized form, due to which, all the calculations were based on normalized value. This lead to a problem to keep feather value for both dimensions and calculate based on that. It also didn't give the smooth solution.
The current implementation resolve this issue by denormalizing the coordinate value and use the Rush type algo without further changes.
TODO: Do we need to handle divide by zero scenario when the feather size is zero?